+++ NEWS.md
@@ -1,3 +1,19 @@
+mt5R v0.1.3 (Release date: 2021-01-30)
+==============
+
+Changes:
+
+* Added `MT5.SymbolExpiration()`.
+* Added `MT5.ServerTime()`.
+* Fixed `MT5.AllSymbols()`.
+* `MT5.GetSymbol()` now accepts `iRows = Inf`.
+* `MT5.MarketIsOpen()` now checks if internet is working as well.
+* `MT5.BidAskBook()` and `MT5.BidAskSpread()` now are adapted to work on auctions.
+* Changed `SUB_REQUEST_DIVISOR` from `!` to `#`. Check [forum](https://github.com/Kinzel/mt5R/discussions/6).
+* Added layers of sanity checks in data-obtain functions.
+* More blocks of texts was translated to english.
+* General bugs fixed and documentation updated.
+
 mt5R v0.1.2 (Release date: 2021-01-16)
 ==============